Curtail strcat() abuses, wave one.
A prevalent antipattern in old parts of TinTin is to repeatedly strcat() the
same string, thus searching for the end of string over and over.  It's usually
enough to do a single sprintf() (!), gaining the possibility of snprintf() for
overflow safety; in the rest of cases we can keep a pointer to the end.